+++ config/environments/test.rb
@@ -64,3 +64,5 @@ Rails.application.configure do
   # Annotate rendered view with file names.
   # config.action_view.annotate_rendered_view_with_filenames = true
 end
+
+Rails.application.default_url_options = Rails.application.config.action_mailer.default_url_options