+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - precalculating of selected by teams ratio for players
 - multiple external ids for games
 - telegram integration
+- notifications layer
 
 ### Modified
 - RefreshSelectedService